refactor: fix every clippy finding without suppressions
34 sites across the workspace, including one Windows-only and one Linux-only
cast found by linting the cross-targets CI builds - both were identity casts
that would have failed a contributor's first `make lint` on those platforms.
No #[allow] was added anywhere.

Several were real clarity wins rather than lint appeasement. Four arguments
that always travelled together through the floor router became a Copper struct,
taking paint and mark from eight arguments to five; Renderer::new and resize
now share a Viewport value, so the logical-versus-physical size split is
expressed once instead of twice; build_bloom_and_groups stopped taking a width
and height that SceneTargets already owned, removing a second source of truth
that could drift; and a rounded -0.7071 literal became FRAC_1_SQRT_2, which is
the exact 45-degree vector its own comment described.

One suggestion was a trap worth recording. Clippy's needless_range_loop on
`for p in 0..28` over a 32-element panel array is a behaviour change: 28 is the
boundary between mosaic and directory panels, and taking the suggestion made an
atlas test fail on seed 0. The range is deliberately narrower than the
collection, which the lint cannot see, so it is now written as
`panels[..28].iter().enumerate()` with the reason stated.

Behaviour is unchanged: 117 tests pass and a deterministic snapshot is
byte-identical before and after.

+++ crates/gibson-app/src/windows.rs
@@ -226,7 +226,7 @@ fn run_fullscreen(settings: Settings) -> Result<(), String> {
                 0,
                 class_name.as_ptr(),
                 null_mut(),
-                (WS_POPUP | WS_VISIBLE) as u32,
+                WS_POPUP | WS_VISIBLE,
                 rc.left,
                 rc.top,
                 w,
